Jenkins: Adjusting the number of points in X-axis
Allowing to adjust the number of points in X-axis of the performance graphs.
The default number of points is set as 99.
Number of points can be changed using OCR_APPS_PLOT_POINTS environment variable.
OCR_APPS_PLOT_POINTS should be a +ve integer value.
For -ve or zero value of OCR_APPS_PLOT_POINTS, the default number of points will be used.
The same change is done in apps.

Change-Id: I36c5b07a5a5e1d90d3ef53b4bb6a47e98a5878e4