OCR: Fix leaking pointer data through ocrXXCreate calls
With the introduction of labeled EDTs/EVTs/DBs, the GUID field
from the create call was copied into the PD message. In many cases,
this value was garbage (if labeled XX was not used). In rare
cases, this value would actually be read and interpreted as an
actual GUID (particularly in the marshalling/unmarshalling code
that is required to either send a message to another PD or to
defer calls).

This simple fixes makes sure that the GUIDs are set to NULL if labeling
is not used.

Change-Id: Ie36fd487e5eb2adf46b6956bbe7aafa5f47ecfd0