OCR TG: Added affinitized scheduler
The scheduler attemps to respect affinities if they
are specified. Otherwise, it works like the other CE scheduler.
Note that this means that there will be twice as many messages
to get work:
  - one saying: "give me affinitized work"
  - one saying: "give mon any work"

This implementation is therefore not as efficient if you do not
specify affinities!! (but it is correct)

Still missing: the way to create affinities and set them properly on TG

Change-Id: Ie8b7582b2837a213181cdfdac28646841269799b