Addition of an allocator monitoring ability in trace.
This includes a altered hc-policy domain to add traces to allocation
and deallocation in the MemAlloc/MemUnalloc and PdMalloc/PdFree of
the hc policy domain.

Also includes the changes necessary in the trace utility to handle
these new trace events.

Needed to add special features to stop recursion in the trace util
itself (it calls PdMalloc) and to handle the trace during shutdown.

Also includes small fix to 'unused' warning when using OCR_TRACE_BINARY.

Change-Id: I0d716b292506cd646d08ed0de48352dd84d19667